requirements.txt
numpy
pandas
accelerate
chromadb
einops
faiss-cpu
fastapi
groq
PyMuPDF
pypdf
tools
pdfplumber
unstructured[pdf]
langchain
langchain-openai
langchain-groq
langchain-community
langchain-chroma
langgraph
langsmith
tiktoken
tavily-python
httpx
huggingface-hub
transformers
torch
torchvision
torchtext
imutils
Jinja2
joblib
numba
nltk
peft
plotly
safetensors
sentence-transformers
sentencepiece
timm
redis
python-dotenv
